+++ lib/src/features/super_select_form_field/presentation/widgets/super_select_form_field.dart
@@ -96,7 +96,7 @@ class SuperSelectFormField<T> extends StatefulWidget {
     this.cursorErrorColor,
     this.style,
     this.strutStyle,
-    this.autovalidateMode = AutovalidateMode.disabled,
+    this.autovalidateMode,
   }) : assert(
          onSaved == null || onSave == null,
          'Provide either onSaved or onSave, not both.',
@@ -237,7 +237,7 @@ class SuperSelectFormField<T> extends StatefulWidget {
   final Color? cursorErrorColor;
   final TextStyle? style;
   final StrutStyle? strutStyle;
-  final AutovalidateMode autovalidateMode;
+  final AutovalidateMode? autovalidateMode;
 
   @override
   State<SuperSelectFormField<T>> createState() =>
@@ -526,7 +526,10 @@ class _SuperSelectFormFieldState<T> extends State<SuperSelectFormField<T>> {
       initialValue: _controller.value,
       enabled: !widget.disabled,
       onSaved: widget.onSaved ?? widget.onSave,
-      autovalidateMode: widget.autovalidateMode,
+      autovalidateMode: SffDecoration.effectiveAutovalidateMode(
+        context,
+        widget.autovalidateMode,
+      ),
       validator: (_) => _controller.error,
       builder: (formState) {
         _controller.configure(